+++ MsgModule/Sources/SettingsManager.h
@@ -15,11 +15,11 @@ namespace sm
 	public:
 		static SettingsManager* Instance();
 		std::wstring GetCurrentNetAddres() const;
-		unsigned long GetCurrentStatesPort() const;
-		unsigned long GetCurrentMessagesPort() const;
+		unsigned short GetCurrentStatesPort() const;
+		unsigned short GetCurrentMessagesPort() const;
 		void SetCurrentAddress(const std::wstring& addr);
-		void SetCurrentStatesPort(unsigned long statePort);
-		void SetCurrentMessagesPort(unsigned long chatPort);
+		void SetCurrentStatesPort(unsigned short statePort);
+		void SetCurrentMessagesPort(unsigned short chatPort);
 		std::vector<std::wstring> GetActiveAddresses() const;
 
 	private:
@@ -32,8 +32,8 @@ namespace sm
 	private:
 		std::unique_ptr<impl::AdaptersAddressHolder> m_addressHolder;
 		std::wstring m_currentNetAddress;
-		unsigned long m_stateMsgPort;
-		unsigned long m_chatMsgPort;
+		unsigned short m_stateMsgPort;
+		unsigned short m_chatMsgPort;
 	};
 }
